remove intermediate versions
